dns: divide implementation in three parts
