+++ src/dns/include.am
@@ -1,12 +1,14 @@
 noinst_LTLIBRARIES += src/dns/libdns.la
 
 src_dns_libdns_la_SOURCES = \
-	src/dns/dns.cpp
+	src/dns/query.cpp \
+	src/dns/resolver.cpp \
+	src/dns/response.cpp
 
 libmeasurement_kit_la_LIBADD += src/dns/libdns.la
 
 dnsincludedir = $(includedir)/measurement_kit/dns
 dnsinclude_HEADERS = \
     include/measurement_kit/dns/query.hpp \
-    include/measurement_kit/dns/response.hpp \
-    include/measurement_kit/dns/resolver.hpp
+    include/measurement_kit/dns/resolver.hpp \
+    include/measurement_kit/dns/response.hpp
